+++ examples/asa/assets/5-contract-asa-stateless.py
@@ -1,10 +1,10 @@
-# Account Contract
+# Contract Account
 
 # Add parent directory to path so that algobpy can be imported
 import sys
 sys.path.insert(0,'..')
 
-from algobpy.parse import parseArgs
+from algobpy.parse import parse_params
 from pyteal import *
 
 def contract_asa(app_id):
@@ -23,7 +23,7 @@ def contract_asa(app_id):
     )
 
     pay = And(
-        # The first transaction must be 
+        # The first transaction must be
         # an ApplicationCall (ie call stateful smart contract)
         Gtxn[0].type_enum() == TxnType.ApplicationCall,
 
@@ -38,7 +38,7 @@ def contract_asa(app_id):
     )
 
     create = And(
-        # The first transaction must be 
+        # The first transaction must be
         # an ApplicationCall (ie call stateful smart contract)
         Gtxn[0].type_enum() == TxnType.ApplicationCall,
 
@@ -58,11 +58,11 @@ def contract_asa(app_id):
 
 if __name__ == "__main__":
     params = {
-        "APP_ID": 1    
+        "APP_ID": 1
     }
 
     # Overwrite params if sys.argv[1] is passed
     if(len(sys.argv) > 1):
-        params = parseArgs(sys.argv[1], params)
+        params = parse_params(sys.argv[1], params)
 
     print(compileTeal(contract_asa(params["APP_ID"]), Mode.Signature))