+++ examples/permissioned-token-freezing/assets/clawback-escrow.py
@@ -4,7 +4,7 @@
 import sys
 sys.path.insert(0,'..')
 
-from algobpy.parse import parseArgs
+from algobpy.parse import parse_params
 from pyteal import *
 
 # source: https://github.com/algorand/smart-contracts/blob/master/devrel/poi/clawback-escrow.teal
@@ -77,6 +77,6 @@ if __name__ == "__main__":
 
     # Overwrite params if sys.argv[1] is passed
     if(len(sys.argv) > 1):
-        params = parseArgs(sys.argv[1], params)
+        params = parse_params(sys.argv[1], params)
 
     print(compileTeal(clawback_escrow(params["ASSET_ID"], params["APP_ID"]), Mode.Signature))